S3 tokens require source and log bucket to be specified (#94)
* S3 tokens require source bucket to be specified

The log bucket is optional, but the source bucket to monitor *must* be specified to make a token. The API will block requests which do not specify the latter.

* S3 log bucket also required to make S3 token.

The console UI generates a random log bucket, but in order to generate via the API this needs to be specified, as it is necessary in order to generate the script for the `/canarytoken/download` request.